requirements.txt
# SPDX-License-Identifier: CC0-1.0
# SPDX-FileCopyrightText: 2022 Harald Sitter <sitter@kde.org>

Appium_Python_Client==3.0.0
Flask==3.0.0
lxml==4.9.3
PyGObject==3.46.0
selenium==4.13.0
